// src/open-api/model/userRole.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type UserRole = 'ADMIN' | 'USER';

export const UserRole = {
    Admin: 'ADMIN' as UserRole,
    User: 'USER' as UserRole
};